Add mysql dump of phabricator DBs to tarsnap backup job
Summary: Add a new vault policy allowing reading DB creds for phabricator DBs. Use this policy for a new task in the backup-tarsnap job that uses mysqldump to dump a SQL file to recreate all the DBs for Phabricator.

Test Plan: Tested by temporarily adding this to the Borg job and running it. It took a few minutes to create the SQL file, since there's a lot of content in the DB, but it was able to do it and back it up without issue.

Reviewers: matt

Reviewed By: matt

Differential Revision: https://code.home.mattmoriarity.com/D3