Add GraphQL API for getting inbox documents from paperless
Summary:
Add a `paperlessInboxDocuments` field to the `Viewer`. It supports basic forward pagination. Auth works by either passing the Cookie header along locally or by passing the X-Auth-Request-User header along when deployed for real.

The code to make the request to Paperless is pretty janky. When I feel like it, I should refactor that to be a well-abstracted client in another package.

Resolves T16.

Test Plan: Tested by running locally and querying with Insomnia.

Reviewers: matt

Reviewed By: matt

Maniphest Tasks: T16

Differential Revision: https://code.home.mattmoriarity.com/D10