Fix display of multi-line commit messages in deploys UI
Summary:
Split the first line of the commit message out as the subject and treat it differently.

In the deploys list, only show the subject. In the detail view, only turn the subject into a link, and if there is more to the message, show that below the subject with linebreaks preserved.

Fixes T14.

Test Plan: Ran locally. Most recent commit messages are formatted this way because they've gone through Phabricator.

Reviewers: matt

Reviewed By: matt

Maniphest Tasks: T14

Differential Revision: https://code.home.mattmoriarity.com/D11