Add Adminer app
Summary: Adds a new Nomad app for Adminer to let me administer all the different databases. Adds an `adminer` vhost to the ingress to access it. There's a custom Adminer plugin that replaces the login form with a selection of pre-configured logins, with credentials populated from Vault on the server.

Test Plan: I installed the app with Nomadic and experimented with it until I got it working.

Reviewers: matt

Reviewed By: matt

Differential Revision: https://code.home.mattmoriarity.com/D21